app/requirements.txt
Flask==0.12.2
Jinja2==2.9.6
google-cloud==0.30.0
google-cloud-datastore==1.4.0